Split dungeon editor state content model
